회원가입/로그인, 마이페이지, 관리자페이지, 포인트 기능 추가 구현
회원가입 및 로그인 시 관리자 아이디 판별, 마이페이지(포인트 내역 조회, 등업 신청), 관리자 로그인, 관리자페이지(탭 메뉴,
체크박스, 회원 목록 조회), 포인트 획득 및 사용 관련 트리거 생성, home.jsp/header.jsp 분리

+++ src/main/java/com/kh/itcom/user/controller/UserController.java
@@ -1,6 +1,7 @@
 package com.kh.itcom.user.controller;
 
 import java.util.HashMap;
+import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
@@ -13,6 +14,9 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.kh.itcom.admin.domain.Admin;
+import com.kh.itcom.admin.service.AdminService;
+import com.kh.itcom.point.domain.PointHistory;
 import com.kh.itcom.user.domain.User;
 import com.kh.itcom.user.service.UserService;
 
@@ -21,6 +25,8 @@ public class UserController {
 
 	@Autowired
 	private UserService uService;
+	@Autowired
+	private AdminService aService;
 
 	/**
 	 * 회원가입 화면
@@ -42,9 +48,16 @@ public class UserController {
 	@RequestMapping(value = "/user/join.do", method = RequestMethod.POST)
 	public ModelAndView userJoin(@ModelAttribute User user, ModelAndView mv) {
 		try {
-			int result = uService.joinUser(user);
-			if (result > 0) {
-				mv.setViewName("redirect:/home.do");
+			int check = aService.checkId(user.getUserId());
+			if (check > 0) {
+				// 사용할 수 없는 아이디입니다.
+			} else {
+				int result = uService.joinUser(user);
+				if (result > 0) {
+					mv.setViewName("redirect:/home.do");
+				} else {
+					// 회원가입 실패
+				}
 			}
 		} catch (Exception e) {
 			System.out.println(e.getMessage()); // error check
@@ -71,13 +84,27 @@ public class UserController {
 	 * @return
 	 */
 	@RequestMapping(value = "/user/login.do", method = RequestMethod.POST)
-	public ModelAndView userLogin(@ModelAttribute User user, HttpServletRequest request, ModelAndView mv) {
+	public ModelAndView userLogin(@RequestParam("loginId") String loginId, @RequestParam("loginPw") String loginPw,
+			HttpServletRequest request, ModelAndView mv) {
 		try {
-			User loginUser = uService.loginUser(user);
-			if (loginUser != null) {
-				HttpSession session = request.getSession();
-				session.setAttribute("loginUser", loginUser);
-				mv.setViewName("redirect:/home.do");
+			int check = aService.checkId(loginId);
+			if (check > 0) {
+				Admin loginAdmin = aService.loginAdmin(new Admin(loginId, loginPw));
+				if (loginAdmin != null) {
+					HttpSession session = request.getSession();
+					session.setAttribute("loginAdmin", loginAdmin);
+					mv.setViewName("redirect:/home.do");
+				}
+			} else {
+				User loginUser = uService.loginUser(new User(loginId, loginPw));
+				if (loginUser != null) {
+					HttpSession session = request.getSession();
+					session.setAttribute("loginUser", loginUser);
+					mv.setViewName("redirect:/home.do");
+				} else {
+					// 로그인 실패
+					System.out.println("회원 로그인 실패");
+				}
 			}
 		} catch (Exception e) {
 			System.out.println(e.getMessage()); // error check
@@ -102,13 +129,23 @@ public class UserController {
 		return mv;
 	}
 
+	/**
+	 * 마이페이지
+	 * 
+	 * @param request
+	 * @param mv
+	 * @return
+	 */
 	@RequestMapping(value = "/user/myPageView.do", method = RequestMethod.GET)
 	public ModelAndView myPageView(HttpServletRequest request, ModelAndView mv) {
 		try {
 			HttpSession session = request.getSession();
 			User user = (User) session.getAttribute("loginUser");
-			mv.addObject("user", user);
-			mv.setViewName("user/myPage");
+			if (user != null) {
+				mv.setViewName("user/myPage");
+			} else {
+				mv.setViewName("redirect:/user/loginView.do");
+			}
 		} catch (Exception e) {
 			System.out.println(e.getMessage()); // error check
 		}
@@ -127,17 +164,20 @@ public class UserController {
 	public ModelAndView userWithdraw(HttpSession session, @RequestParam("userPw") String userPw, ModelAndView mv) {
 		try {
 			User user = (User) session.getAttribute("loginUser");
-			String userId = user.getUserId();
-			HashMap<String, String> userInfo = new HashMap<String, String>() {
-				{
-					put("userId", userId);
-					put("userPw", userPw);
+			if (user != null) {
+				String userId = user.getUserId();
+				HashMap<String, String> userInfo = new HashMap<String, String>() {
+					{
+						put("userId", userId);
+						put("userPw", userPw);
+					}
+				};
+				int result = uService.removeUser(userInfo);
+				if (result > 0) {
+					mv.setViewName("redirect:/user/logout.do");
 				}
-			};
-			System.out.println(userInfo); // value check
-			int result = uService.removeUser(userInfo);
-			if (result > 0) {
-				mv.setViewName("redirect:/user/logout.do");
+			} else {
+				mv.setViewName("redirect:/user/loginView.do");
 			}
 		} catch (Exception e) {
 			System.out.println(e.toString()); // error check
